NodePool CLI: change NodePool CLI to have platform sub command
Change the NodePool CLI to be platform aware
This change is done as part of the multy platform support,
to allow the CLI to provide platform specific UX where needed
